requirements.txt
bidict==0.21.3; python_version >= "3.6"
black==21.9b0; python_full_version >= "3.6.2"
click==8.0.1; python_version >= "3.6" and python_full_version >= "3.6.2"
colorama==0.4.4; python_version >= "3.6" and python_full_version >= "3.6.2" and platform_system == "Windows"
dnspython==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
eventlet==0.30.2
greenlet==1.1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
gunicorn==20.1.0; python_version >= "3.5"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
pathspec==0.9.0; python_full_version >= "3.6.2"
platformdirs==2.3.0; python_version >= "3.6" and python_full_version >= "3.6.2"
python-engineio==4.2.1; python_version >= "3.6"
python-socketio==5.4.0; python_version >= "3.6"
regex==2021.8.28; python_full_version >= "3.6.2"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
tomli==1.2.1; python_version >= "3.6" and python_full_version >= "3.6.2"
typing-extensions==3.10.0.2
